backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.4",
    "fastembed==0.7.4",
    "google-generativeai>=0.8.5",
    "pydantic==2.9.2",
    "python-dotenv==1.0.1",
    "python-multipart==0.0.20",
    "qdrant-client>=1.10.0",
    "uvicorn==0.32.0",
]
